requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
Markdown==3.4.1
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
